requirements-macos.txt
absl-py==2.3.1
antlr4-python3-runtime==4.13.2
astunparse==1.6.3
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
coverage==7.10.6
coveralls==4.0.1
crc32c==2.7.1
cudaq==0.11.0
cycler==0.12.1
dill==0.4.0
distlib==0.4.0
docopt==0.6.2
donfig==0.8.1.post1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
gurobipy==11.0.3
h5py==3.14.0
identify==2.6.14
idna==3.10
iniconfig==2.1.0
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.5.3
mpi4py==4.1.0
namex==0.1.0
networkx==3.5
nodeenv==1.9.1
numcodecs==0.16.2
numpy==2.3.2
openqasm3==1.0.1
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==4.3.0
protobuf==6.32.0
psutil==7.0.0
pydot==4.0.1
pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.3
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.2
qiskit==2.1.2
qiskit-aer==0.17.1
qiskit-qasm3-import==0.6.0
requests==2.32.5
rich==14.1.0
rustworkx==0.17.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
stevedore==5.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.15.0
urllib3==2.5.0
virtualenv==20.34.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
zarr==3.1.2
